Fix SAF display issues
With previous criteria, SAF would be removed from display when there
were no reports and SAF itself wasn't used. This becomes a problem when
user creates a query which returns no reports and then manually deletes
the query. The action satisfied both requirements and SAF would be
removed from display, leaving no way to fix the issue but to refresh
the page.

Now, instead of using the mentioned requirements, a request is made to
the server to check if ANY reports exist - only if there are none, SAF
is removed from display (which is how it was initially intended to work).